# Improve Go speed