Faster Go code by interleaving operations and inling zipper code.

```
benchmark                    old ns/op     new ns/op     delta
BenchmarkWrite_8-32          9.24          8.66          -6.28%
BenchmarkWrite_16-32         15.3          13.8          -9.78%
BenchmarkWrite_64-32         44.6          38.9          -12.84%
BenchmarkWrite_1K-32         668           564           -15.66%
BenchmarkWrite_8K-32         5252          4495          -14.41%
BenchmarkSum64_8-32          146           128           -12.05%
BenchmarkSum64_16-32         148           132           -10.74%
BenchmarkSum64_64-32         154           136           -12.18%
BenchmarkSum64_1K-32         776           658           -15.30%
BenchmarkSum64_8K-32         5376          4616          -14.14%
BenchmarkSum256_8-32         300           263           -12.40%
BenchmarkSum256_16-32        304           266           -12.47%
BenchmarkSum256_64-32        320           263           -17.88%
BenchmarkSum256_1K-32        939           794           -15.47%
BenchmarkSum256_8K-32        5548          4712          -15.07%
BenchmarkSum256_1M-32        681705        575480        -15.58%
BenchmarkSum256_5M-32        3409493       2876752       -15.63%
BenchmarkSum256_10M-32       6808693       5759719       -15.41%
BenchmarkSum256_25M-32       17062713      14422506      -15.47%
BenchmarkParallel_1M-32      44241         37559         -15.10%
BenchmarkParallel_5M-32      220972        188019        -14.91%
BenchmarkParallel_10M-32     454317        377080        -17.00%
BenchmarkParallel_25M-32     1111869       947157        -14.81%

benchmark                    old MB/s     new MB/s     speedup
BenchmarkWrite_8-32          866.18       924.18       1.07x
BenchmarkWrite_16-32         1043.84      1157.03      1.11x
BenchmarkWrite_64-32         1434.39      1645.50      1.15x
BenchmarkWrite_1K-32         1532.63      1817.21      1.19x
BenchmarkWrite_8K-32         1559.83      1822.28      1.17x
BenchmarkSum64_8-32          54.81        62.31        1.14x
BenchmarkSum64_16-32         108.03       121.03       1.12x
BenchmarkSum64_64-32         414.85       472.26       1.14x
BenchmarkSum64_1K-32         1319.03      1557.48      1.18x
BenchmarkSum64_8K-32         1523.91      1774.69      1.16x
BenchmarkSum256_8-32         26.66        30.43        1.14x
BenchmarkSum256_16-32        52.65        60.16        1.14x
BenchmarkSum256_64-32        199.76       243.26       1.22x
BenchmarkSum256_1K-32        1090.84      1290.44      1.18x
BenchmarkSum256_8K-32        1476.49      1738.54      1.18x
BenchmarkSum256_1M-32        1538.17      1822.09      1.18x
BenchmarkSum256_5M-32        1537.73      1822.50      1.19x
BenchmarkSum256_10M-32       1540.05      1820.53      1.18x
BenchmarkSum256_25M-32       1536.36      1817.60      1.18x
BenchmarkParallel_1M-32      23701.69     27917.92     1.18x
BenchmarkParallel_5M-32      23726.40     27884.84     1.18x
BenchmarkParallel_10M-32     23080.26     27807.81     1.20x
BenchmarkParallel_25M-32     23576.87     27676.93     1.17x
```